Started code to make all permutations of oligos for a given pos
Added RNASequence class, and wrote code to generate all different levels of all given codons for the sequence. Still needs to link them and check overlaps, as well as fill in the rest of the codons for the given codon's corresponding acid. Also has no form of input other than hard-coding. Comments and logical ordering of code to come later.